fix assets.d.ts
