+++ src/renderer/assets.d.ts
@@ -1,15 +1,33 @@
-declare module "*.png" {
-  const src: string;
-  export default src;
-}
+/* eslint-disable import/no-default-export */
+type Styles = Record<string, string>;
 
 declare module "*.svg" {
-  import type * as React from "react";
+  export const ReactComponent: React.FC<React.SVGAttributes<SVGElement>>;
+  const content: string;
+  export default content;
+}
 
-  // For `import { ReactComponent as Icon } from "./icon.svg"`
-  export const ReactComponent: React.FC<React.SVGProps<SVGSVGElement> & { title?: string }>;
+declare module "*.png" {
+  const content: string;
+  export default content;
+}
 
-  // For `import iconUrl from "./icon.svg"`
-  const src: string;
-  export default src;
+declare module "*.jpg" {
+  const content: string;
+  export default content;
+}
+
+declare module "*.scss" {
+  const content: Styles;
+  export default content;
+}
+
+declare module "*.sass" {
+  const content: Styles;
+  export default content;
+}
+
+declare module "*.css" {
+  const content: Styles;
+  export default content;
 }
